pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pizza-order-extractor"
version = "0.1.0"
description = "Extract order list from Google Spreadsheet"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
keywords = ["google-sheets", "orders", "spreadsheet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.48.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/pizza"
Repository = "https://github.com/yourusername/pizza.git"
Issues = "https://github.com/yourusername/pizza/issues"

[project.scripts]
extract-orders = "extract_orders:main"
place-orders = "place_orders:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.setuptools]
py-modules = ["extract_orders", "place_orders"]
